Overhaul project
- Restructured packages and class names
- Implemented config handler to control FOV behavior and reloading of block configs
- Tweaked loading logic and logging to skip and no longer print errors when specified mods are not loaded
- Added some useful logging
- Minor code clean-up
- Bumped version to 1.7.0
