+++ gradle.properties
@@ -15,7 +15,7 @@ show_testing_output = false
 
 # Mod Information
 # HIGHLY RECOMMEND complying with SemVer for mod_version: https://semver.org/
-mod_version = 1.6.1
+mod_version = 1.7.0
 root_package = tfar.paths
 mod_id = paths
 mod_name = Paths
