--- src/main/java/com/tfar/paths/ExampleConfig.java
@@ -1,543 +1,0 @@
-package com.tfar.paths;
-
-public class ExampleConfig {
-  static String s = "{\n" +
-          "\"0.5,4,false\": {\n" +
-          "\"minecraft:nether_wart_block\": [],\n" +
-          "\"minecraft:soul_sand\":[],\n" +
-          "\"minecraft:slime\": [],\n" +
-          "\"minecraft:tnt\": [],\n" +
-          "\"tfc:peat_grass\":[]\n" +
-          "},\n" +
-          "\"0.75,3,false\": {\n" +
-          "\"minecraft:dirt\": [2],\n" +
-          "\"minecraft:mycelium\":[],\n" +
-          "\"minecraft:web\": [],\n" +
-          "\"tfc:peat\":[]\n" +
-          "},\n" +
-          "\"1.33,2,true\": {\n" +
-          "\"minecraft:grass_path\": [],\n" +
-          "\"minecraft:plank\": [],\n" +
-          "\"minecraft:oak_stairs\": [],\n" +
-          "\"minecraft:spruce_stairs\": [],\n" +
-          "\"minecraft:birch_stairs\": [],\n" +
-          "\"minecraft:jungle_stairs\": [],\n" +
-          "\"minecraft:acacia_stairs\": [],\n" +
-          "\"minecraft:dark_oak_stairs\": [],\n" +
-          "\"minecraft:wooden_slab\": [],\n" +
-          "\"minecraft:double_wooden_slab\": [],\n" +
-          "\"minecraft:stone_stairs\": [],\n" +
-          "\"minecraft:cobblestone\": [],\n" +
-          "\"minecraft:stone\": [2,4,6],\n" +
-          "\"minecraft:stone_slab\": [],\n" +
-          "\"minecraft:double_stone_slab\": [],\n" +
-          "\"minecraft:stonebrick\": [0,3],\n" +
-          "\"minecraft:stone_brick_stairs\": [],\n" +
-          "\"minecraft:sandstone\": [1,2],\n" +
-          "\"minecraft:sandstone_stairs\": [],\n" +
-          "\"minecraft:red_sandstone\": [1,2],\n" +
-          "\"minecraft:red_sandstone_stairs\": [],\n" +
-          "\"minecraft:stone_slab2\": [],\n" +
-          "\"minecraft:brick_block\": [],\n" +
-          "\"minecraft:brick_stairs\": [],\n" +
-          "\"minecraft:glass\": [],\n" +
-          "\"minecraft:white_glazed_terracotta\": [],\n" +
-          "\"minecraft:orange_glazed_terracotta\": [],\n" +
-          "\"minecraft:magenta_glazed_terracotta\": [],\n" +
-          "\"minecraft:light_blue_glazed_terracotta\": [],\n" +
-          "\"minecraft:yellow_glazed_terracotta\": [],\n" +
-          "\"minecraft:lime_glazed_terracotta\": [],\n" +
-          "\"minecraft:pink_glazed_terracotta\": [],\n" +
-          "\"minecraft:gray_glazed_terracotta\": [],\n" +
-          "\"minecraft:silver_glazed_terracotta\": [],\n" +
-          "\"minecraft:cyan_glazed_terracotta\": [],\n" +
-          "\"minecraft:purple_glazed_terracotta\": [],\n" +
-          "\"minecraft:blue_glazed_terracotta\": [],\n" +
-          "\"minecraft:brown_glazed_terracotta\": [],\n" +
-          "\"minecraft:green_glazed_terracotta\": [],\n" +
-          "\"minecraft:red_glazed_terracotta\": [],\n" +
-          "\"minecraft:black_glazed_terracotta\": [],\n" +
-          "\"minecraft:endstone\": [],\n" +
-          "\"minecraft:end_bricks\": [],\n" +
-          "\"tfc:path/andesite\": [],\n" +
-          "\"tfc:path/basalt\": [],\n" +
-          "\"tfc:path/chalk\": [],\n" +
-          "\"tfc:path/chert\": [],\n" +
-          "\"tfc:path/claystone\": [],\n" +
-          "\"tfc:path/conglomerate\": [],\n" +
-          "\"tfc:path/dacite\": [],\n" +
-          "\"tfc:path/diorite\": [],\n" +
-          "\"tfc:path/dolomite\": [],\n" +
-          "\"tfc:path/gabbro\": [],\n" +
-          "\"tfc:path/gneiss\": [],\n" +
-          "\"tfc:path/granite\": [],\n" +
-          "\"tfc:path/limestone\": [],\n" +
-          "\"tfc:path/marble\": [],\n" +
-          "\"tfc:path/phyllite\": [],\n" +
-          "\"tfc:path/quartzite\": [],\n" +
-          "\"tfc:path/rhyolite\": [],\n" +
-          "\"tfc:path/rocksalt\": [],\n" +
-          "\"tfc:path/schist\": [],\n" +
-          "\"tfc:path/shale\": [],\n" +
-          "\"tfc:path/slate\": [],\n" +
-          "\"tfc:wood/log/acacia\": [],\n" +
-          "\"tfc:wood/log/ash\": [],\n" +
-          "\"tfc:wood/log/aspen\": [],\n" +
-          "\"tfc:wood/log/birch\": [],\n" +
-          "\"tfc:wood/log/blackwood\": [],\n" +
-          "\"tfc:wood/log/chestnut\": [],\n" +
-          "\"tfc:wood/log/douglas_fir\": [],\n" +
-          "\"tfc:wood/log/hickory\": [],\n" +
-          "\"tfc:wood/log/kapok\": [],\n" +
-          "\"tfc:wood/log/maple\": [],\n" +
-          "\"tfc:wood/log/oak\": [],\n" +
-          "\"tfc:wood/log/palm\": [],\n" +
-          "\"tfc:wood/log/pine\": [],\n" +
-          "\"tfc:wood/log/rosewood\": [],\n" +
-          "\"tfc:wood/log/sequoia\": [],\n" +
-          "\"tfc:wood/log/spruce\": [],\n" +
-          "\"tfc:wood/log/sycamore\": [],\n" +
-          "\"tfc:wood/log/white_cedar\": [],\n" +
-          "\"tfc:wood/log/willow\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_acacia\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_ash\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_aspen\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_birch\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_blackwood\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_chestnut\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_douglas_fir\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_hickory\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_kapok\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_maple\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_oak\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_palm\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_pine\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_rosewood\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_sequoia\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_spruce\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_sycamore\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_white_cedar\": [],\n" +
-          "\"debark:debarked_log_tfc_wood_log_willow\": [],\n" +
-          "\"tfc:log_pile\": [],\n" +
-          "\"tfc:wood/planks/acacia\": [],\n" +
-          "\"tfc:wood/planks/ash\": [],\n" +
-          "\"tfc:wood/planks/aspen\": [],\n" +
-          "\"tfc:wood/planks/birch\": [],\n" +
-          "\"tfc:wood/planks/blackwood\": [],\n" +
-          "\"tfc:wood/planks/chestnut\": [],\n" +
-          "\"tfc:wood/planks/douglas_fir\": [],\n" +
-          "\"tfc:wood/planks/hickory\": [],\n" +
-          "\"tfc:wood/planks/kapok\": [],\n" +
-          "\"tfc:wood/planks/maple\": [],\n" +
-          "\"tfc:wood/planks/oak\": [],\n" +
-          "\"tfc:wood/planks/palm\": [],\n" +
-          "\"tfc:wood/planks/pine\": [],\n" +
-          "\"tfc:wood/planks/rosewood\": [],\n" +
-          "\"tfc:wood/planks/sequoia\": [],\n" +
-          "\"tfc:wood/planks/spruce\": [],\n" +
-          "\"tfc:wood/planks/sycamore\": [],\n" +
-          "\"tfc:wood/planks/white_cedar\": [],\n" +
-          "\"tfc:wood/planks/willow\": [],\n" +
-          "\"tfc:stairs/wood/acacia\": [],\n" +
-          "\"tfc:stairs/wood/ash\": [],\n" +
-          "\"tfc:stairs/wood/aspen\": [],\n" +
-          "\"tfc:stairs/wood/birch\": [],\n" +
-          "\"tfc:stairs/wood/blackwood\": [],\n" +
-          "\"tfc:stairs/wood/chestnut\": [],\n" +
-          "\"tfc:stairs/wood/douglas_fir\": [],\n" +
-          "\"tfc:stairs/wood/hickory\": [],\n" +
-          "\"tfc:stairs/wood/kapok\": [],\n" +
-          "\"tfc:stairs/wood/maple\": [],\n" +
-          "\"tfc:stairs/wood/oak\": [],\n" +
-          "\"tfc:stairs/wood/palm\": [],\n" +
-          "\"tfc:stairs/wood/pine\": [],\n" +
-          "\"tfc:stairs/wood/rosewood\": [],\n" +
-          "\"tfc:stairs/wood/sequoia\": [],\n" +
-          "\"tfc:stairs/wood/spruce\": [],\n" +
-          "\"tfc:stairs/wood/sycamore\": [],\n" +
-          "\"tfc:stairs/wood/white_cedar\": [],\n" +
-          "\"tfc:stairs/wood/willow\": [],\n" +
-          "\"tfc:slab/wood/acacia\": [],\n" +
-          "\"tfc:slab/wood/ash\": [],\n" +
-          "\"tfc:slab/wood/aspen\": [],\n" +
-          "\"tfc:slab/wood/birch\": [],\n" +
-          "\"tfc:slab/wood/blackwood\": [],\n" +
-          "\"tfc:slab/wood/chestnut\": [],\n" +
-          "\"tfc:slab/wood/douglas_fir\": [],\n" +
-          "\"tfc:slab/wood/hickory\": [],\n" +
-          "\"tfc:slab/wood/kapok\": [],\n" +
-          "\"tfc:slab/wood/maple\": [],\n" +
-          "\"tfc:slab/wood/oak\": [],\n" +
-          "\"tfc:slab/wood/palm\": [],\n" +
-          "\"tfc:slab/wood/pine\": [],\n" +
-          "\"tfc:slab/wood/rosewood\": [],\n" +
-          "\"tfc:slab/wood/sequoia\": [],\n" +
-          "\"tfc:slab/wood/spruce\": [],\n" +
-          "\"tfc:slab/wood/sycamore\": [],\n" +
-          "\"tfc:slab/wood/white_cedar\": [],\n" +
-          "\"tfc:slab/wood/willow\": [],\n" +
-          "\"tfc:double_slab/wood/acacia\": [],\n" +
-          "\"tfc:double_slab/wood/ash\": [],\n" +
-          "\"tfc:double_slab/wood/aspen\": [],\n" +
-          "\"tfc:double_slab/wood/birch\": [],\n" +
-          "\"tfc:double_slab/wood/blackwood\": [],\n" +
-          "\"tfc:double_slab/wood/chestnut\": [],\n" +
-          "\"tfc:double_slab/wood/douglas_fir\": [],\n" +
-          "\"tfc:double_slab/wood/hickory\": [],\n" +
-          "\"tfc:double_slab/wood/kapok\": [],\n" +
-          "\"tfc:double_slab/wood/maple\": [],\n" +
-          "\"tfc:double_slab/wood/oak\": [],\n" +
-          "\"tfc:double_slab/wood/palm\": [],\n" +
-          "\"tfc:double_slab/wood/pine\": [],\n" +
-          "\"tfc:double_slab/wood/rosewood\": [],\n" +
-          "\"tfc:double_slab/wood/sequoia\": [],\n" +
-          "\"tfc:double_slab/wood/spruce\": [],\n" +
-          "\"tfc:double_slab/wood/sycamore\": [],\n" +
-          "\"tfc:double_slab/wood/white_cedar\": [],\n" +
-          "\"tfc:double_slab/wood/willow\": [],\n" +
-          "\"tfc:cobble/granite\": [],\n" +
-          "\"tfc:cobble/diorite\": [],\n" +
-          "\"tfc:cobble/gabbro\": [],\n" +
-          "\"tfc:cobble/shale\": [],\n" +
-          "\"tfc:cobble/claystone\": [],\n" +
-          "\"tfc:cobble/rocksalt\": [],\n" +
-          "\"tfc:cobble/limestone\": [],\n" +
-          "\"tfc:cobble/conglomerate\": [],\n" +
-          "\"tfc:cobble/dolomite\": [],\n" +
-          "\"tfc:cobble/chert\": [],\n" +
-          "\"tfc:cobble/chalk\": [],\n" +
-          "\"tfc:cobble/rhyolite\": [],\n" +
-          "\"tfc:cobble/basalt\": [],\n" +
-          "\"tfc:cobble/andesite\": [],\n" +
-          "\"tfc:cobble/dacite\": [],\n" +
-          "\"tfc:cobble/quartzite\": [],\n" +
-          "\"tfc:cobble/slate\": [],\n" +
-          "\"tfc:cobble/phyllite\": [],\n" +
-          "\"tfc:cobble/schist\": [],\n" +
-          "\"tfc:cobble/gneiss\": [],\n" +
-          "\"tfc:cobble/marble\": [],\n" +
-          "\"tfc:stairs/cobble/andesite\": [],\n" +
-          "\"tfc:stairs/cobble/basalt\": [],\n" +
-          "\"tfc:stairs/cobble/chalk\": [],\n" +
-          "\"tfc:stairs/cobble/chert\": [],\n" +
-          "\"tfc:stairs/cobble/claystone\": [],\n" +
-          "\"tfc:stairs/cobble/conglomerate\": [],\n" +
-          "\"tfc:stairs/cobble/dacite\": [],\n" +
-          "\"tfc:stairs/cobble/diorite\": [],\n" +
-          "\"tfc:stairs/cobble/dolomite\": [],\n" +
-          "\"tfc:stairs/cobble/gabbro\": [],\n" +
-          "\"tfc:stairs/cobble/gneiss\": [],\n" +
-          "\"tfc:stairs/cobble/granite\": [],\n" +
-          "\"tfc:stairs/cobble/limestone\": [],\n" +
-          "\"tfc:stairs/cobble/marble\": [],\n" +
-          "\"tfc:stairs/cobble/phyllite\": [],\n" +
-          "\"tfc:stairs/cobble/quartzite\": [],\n" +
-          "\"tfc:stairs/cobble/rhyolite\": [],\n" +
-          "\"tfc:stairs/cobble/rocksalt\": [],\n" +
-          "\"tfc:stairs/cobble/schist\": [],\n" +
-          "\"tfc:stairs/cobble/shale\": [],\n" +
-          "\"tfc:stairs/cobble/slate\": [],\n" +
-          "\"tfc:slab/cobble/andesite\": [],\n" +
-          "\"tfc:slab/cobble/basalt\": [],\n" +
-          "\"tfc:slab/cobble/chalk\": [],\n" +
-          "\"tfc:slab/cobble/chert\": [],\n" +
-          "\"tfc:slab/cobble/claystone\": [],\n" +
-          "\"tfc:slab/cobble/conglomerate\": [],\n" +
-          "\"tfc:slab/cobble/dacite\": [],\n" +
-          "\"tfc:slab/cobble/diorite\": [],\n" +
-          "\"tfc:slab/cobble/dolomite\": [],\n" +
-          "\"tfc:slab/cobble/gabbro\": [],\n" +
-          "\"tfc:slab/cobble/gneiss\": [],\n" +
-          "\"tfc:slab/cobble/granite\": [],\n" +
-          "\"tfc:slab/cobble/limestone\": [],\n" +
-          "\"tfc:slab/cobble/marble\": [],\n" +
-          "\"tfc:slab/cobble/phyllite\": [],\n" +
-          "\"tfc:slab/cobble/quartzite\": [],\n" +
-          "\"tfc:slab/cobble/rhyolite\": [],\n" +
-          "\"tfc:slab/cobble/rocksalt\": [],\n" +
-          "\"tfc:slab/cobble/schist\": [],\n" +
-          "\"tfc:slab/cobble/shale\": [],\n" +
-          "\"tfc:slab/cobble/slate\": [],\n" +
-          "\"tfc:double_slab/cobble/andesite\": [],\n" +
-          "\"tfc:double_slab/cobble/basalt\": [],\n" +
-          "\"tfc:double_slab/cobble/chalk\": [],\n" +
-          "\"tfc:double_slab/cobble/chert\": [],\n" +
-          "\"tfc:double_slab/cobble/claystone\": [],\n" +
-          "\"tfc:double_slab/cobble/conglomerate\": [],\n" +
-          "\"tfc:double_slab/cobble/dacite\": [],\n" +
-          "\"tfc:double_slab/cobble/diorite\": [],\n" +
-          "\"tfc:double_slab/cobble/dolomite\": [],\n" +
-          "\"tfc:double_slab/cobble/gabbro\": [],\n" +
-          "\"tfc:double_slab/cobble/gneiss\": [],\n" +
-          "\"tfc:double_slab/cobble/granite\": [],\n" +
-          "\"tfc:double_slab/cobble/limestone\": [],\n" +
-          "\"tfc:double_slab/cobble/marble\": [],\n" +
-          "\"tfc:double_slab/cobble/phyllite\": [],\n" +
-          "\"tfc:double_slab/cobble/quartzite\": [],\n" +
-          "\"tfc:double_slab/cobble/rhyolite\": [],\n" +
-          "\"tfc:double_slab/cobble/rocksalt\": [],\n" +
-          "\"tfc:double_slab/cobble/schist\": [],\n" +
-          "\"tfc:double_slab/cobble/shale\": [],\n" +
-          "\"tfc:double_slab/cobble/slate\": []\n" +
-          "},\n" +
-          "\"1.5,2,true\": {\n" +
-          "\"minecraft:crafting_table\": [],\n" +
-          "\"minecraft:furnace\": [],\n" +
-          "\"minecraft:dispenser\": [],\n" +
-          "\"minecraft:dropper\": [],\n" +
-          "\"minecraft:log\": [],\n" +
-          "\"minecraft:log2\": [],\n" +
-          "\"debark:debarked_log_minecraft_log\": [],\n" +
-          "\"debark:debarked_log_minecraft_log2\": [],\n" +
-          "\"minecraft:wool\": [],\n" +
-          "\"minecraft:carpet\": [],\n" +
-          "\"minecraft:concrete\": [],\n" +
-          "\"minecraft:mossy_cobblestone\": [],\n" +
-          "\"minecraft:stonebrick\": [1,2],\n" +
-          "\"minecraft:stone_slab\": [6,7],\n" +
-          "\"minecraft:purpur_block\": [],\n" +
-          "\"minecraft:purpur_pillar\": [],\n" +
-          "\"minecraft:purpur_slab\": [],\n" +
-          "\"minecraft:nether_brick_stairs\": [],\n" +
-          "\"minecraft:quartz_stairs\": [],\n" +
-          "\"minecraft:purpur_stairs\": [],\n" +
-          "\"minecraft:quartz_block\": [],\n" +
-          "\"minecraft:coal_block\": [],\n" +
-          "\"minecraft:redstone_block\": [],\n" +
-          "\"minecraft:nether_brick\": [],\n" +
-          "\"minecraft:prismarine\": [],\n" +
-          "\"minecraft:sea_lantern\": [],\n" +
-          "\"minecraft:packed_ice\": [],\n" +
-          "\"minecraft:bone_block\": [],\n" +
-          "\"tfc:wood/workbench/acacia\": [],\n" +
-          "\"tfc:wood/workbench/ash\": [],\n" +
-          "\"tfc:wood/workbench/aspen\": [],\n" +
-          "\"tfc:wood/workbench/birch\": [],\n" +
-          "\"tfc:wood/workbench/blackworkbench\": [],\n" +
-          "\"tfc:wood/workbench/chestnut\": [],\n" +
-          "\"tfc:wood/workbench/douglas_fir\": [],\n" +
-          "\"tfc:wood/workbench/hickory\": [],\n" +
-          "\"tfc:wood/workbench/kapok\": [],\n" +
-          "\"tfc:wood/workbench/maple\": [],\n" +
-          "\"tfc:wood/workbench/oak\": [],\n" +
-          "\"tfc:wood/workbench/palm\": [],\n" +
-          "\"tfc:wood/workbench/pine\": [],\n" +
-          "\"tfc:wood/workbench/roseworkbench\": [],\n" +
-          "\"tfc:wood/workbench/sequoia\": [],\n" +
-          "\"tfc:wood/workbench/spruce\": [],\n" +
-          "\"tfc:wood/workbench/sycamore\": [],\n" +
-          "\"tfc:wood/workbench/white_cedar\": [],\n" +
-          "\"tfc:wood/workbench/willow\": [],\n" +
-          "\"tfc:bricks/andesite\": [],\n" +
-          "\"tfc:bricks/basalt\": [],\n" +
-          "\"tfc:bricks/chalk\": [],\n" +
-          "\"tfc:bricks/chert\": [],\n" +
-          "\"tfc:bricks/claystone\": [],\n" +
-          "\"tfc:bricks/conglomerate\": [],\n" +
-          "\"tfc:bricks/dacite\": [],\n" +
-          "\"tfc:bricks/diorite\": [],\n" +
-          "\"tfc:bricks/dolomite\": [],\n" +
-          "\"tfc:bricks/gabbro\": [],\n" +
-          "\"tfc:bricks/gneiss\": [],\n" +
-          "\"tfc:bricks/granite\": [],\n" +
-          "\"tfc:bricks/limestone\": [],\n" +
-          "\"tfc:bricks/marble\": [],\n" +
-          "\"tfc:bricks/phyllite\": [],\n" +
-          "\"tfc:bricks/quartzite\": [],\n" +
-          "\"tfc:bricks/rhyolite\": [],\n" +
-          "\"tfc:bricks/rocksalt\": [],\n" +
-          "\"tfc:bricks/schist\": [],\n" +
-          "\"tfc:bricks/shale\": [],\n" +
-          "\"tfc:bricks/slate\": [],\n" +
-          "\"tfc:stairs/bricks/andesite\": [],\n" +
-          "\"tfc:stairs/bricks/basalt\": [],\n" +
-          "\"tfc:stairs/bricks/chalk\": [],\n" +
-          "\"tfc:stairs/bricks/chert\": [],\n" +
-          "\"tfc:stairs/bricks/claystone\": [],\n" +
-          "\"tfc:stairs/bricks/conglomerate\": [],\n" +
-          "\"tfc:stairs/bricks/dacite\": [],\n" +
-          "\"tfc:stairs/bricks/diorite\": [],\n" +
-          "\"tfc:stairs/bricks/dolomite\": [],\n" +
-          "\"tfc:stairs/bricks/gabbro\": [],\n" +
-          "\"tfc:stairs/bricks/gneiss\": [],\n" +
-          "\"tfc:stairs/bricks/granite\": [],\n" +
-          "\"tfc:stairs/bricks/limestone\": [],\n" +
-          "\"tfc:stairs/bricks/marble\": [],\n" +
-          "\"tfc:stairs/bricks/phyllite\": [],\n" +
-          "\"tfc:stairs/bricks/quartzite\": [],\n" +
-          "\"tfc:stairs/bricks/rhyolite\": [],\n" +
-          "\"tfc:stairs/bricks/rocksalt\": [],\n" +
-          "\"tfc:stairs/bricks/schist\": [],\n" +
-          "\"tfc:stairs/bricks/shale\": [],\n" +
-          "\"tfc:stairs/bricks/slate\": [],\n" +
-          "\"tfc:slab/bricks/andesite\": [],\n" +
-          "\"tfc:slab/bricks/basalt\": [],\n" +
-          "\"tfc:slab/bricks/chalk\": [],\n" +
-          "\"tfc:slab/bricks/chert\": [],\n" +
-          "\"tfc:slab/bricks/claystone\": [],\n" +
-          "\"tfc:slab/bricks/conglomerate\": [],\n" +
-          "\"tfc:slab/bricks/dacite\": [],\n" +
-          "\"tfc:slab/bricks/diorite\": [],\n" +
-          "\"tfc:slab/bricks/dolomite\": [],\n" +
-          "\"tfc:slab/bricks/gabbro\": [],\n" +
-          "\"tfc:slab/bricks/gneiss\": [],\n" +
-          "\"tfc:slab/bricks/granite\": [],\n" +
-          "\"tfc:slab/bricks/limestone\": [],\n" +
-          "\"tfc:slab/bricks/marble\": [],\n" +
-          "\"tfc:slab/bricks/phyllite\": [],\n" +
-          "\"tfc:slab/bricks/quartzite\": [],\n" +
-          "\"tfc:slab/bricks/rhyolite\": [],\n" +
-          "\"tfc:slab/bricks/rocksalt\": [],\n" +
-          "\"tfc:slab/bricks/schist\": [],\n" +
-          "\"tfc:slab/bricks/shale\": [],\n" +
-          "\"tfc:slab/bricks/slate\": [],\n" +
-          "\"tfc:double_slab/bricks/andesite\": [],\n" +
-          "\"tfc:double_slab/bricks/basalt\": [],\n" +
-          "\"tfc:double_slab/bricks/chalk\": [],\n" +
-          "\"tfc:double_slab/bricks/chert\": [],\n" +
-          "\"tfc:double_slab/bricks/claystone\": [],\n" +
-          "\"tfc:double_slab/bricks/conglomerate\": [],\n" +
-          "\"tfc:double_slab/bricks/dacite\": [],\n" +
-          "\"tfc:double_slab/bricks/diorite\": [],\n" +
-          "\"tfc:double_slab/bricks/dolomite\": [],\n" +
-          "\"tfc:double_slab/bricks/gabbro\": [],\n" +
-          "\"tfc:double_slab/bricks/gneiss\": [],\n" +
-          "\"tfc:double_slab/bricks/granite\": [],\n" +
-          "\"tfc:double_slab/bricks/limestone\": [],\n" +
-          "\"tfc:double_slab/bricks/marble\": [],\n" +
-          "\"tfc:double_slab/bricks/phyllite\": [],\n" +
-          "\"tfc:double_slab/bricks/quartzite\": [],\n" +
-          "\"tfc:double_slab/bricks/rhyolite\": [],\n" +
-          "\"tfc:double_slab/bricks/rocksalt\": [],\n" +
-          "\"tfc:double_slab/bricks/schist\": [],\n" +
-          "\"tfc:double_slab/bricks/shale\": [],\n" +
-          "\"tfc:double_slab/bricks/slate\": []\n" +
-          "},\n" +
-          "\"1.67,2,false\": {\n" +
-          "\"minecraft:hopper\": [],\n" +
-          "\"minecraft:lapis_block\": [],\n" +
-          "\"minecraft:iron_block\": [],\n" +
-          "\"minecraft:emerald_block\": [],\n" +
-          "\"minecraft:glowstone\": [],\n" +
-          "\"minecraft:redstone_lamp\": [],\n" +
-          "\"minecraft:lit_redstone_lamp\": [],\n" +
-          "\"minecraft:sponge\": [],\n" +
-          "\"minecraft:red_nether_brick\": [],\n" +
-          "\"minecraft:white_shulker_box\": [],\n" +
-          "\"minecraft:orange_shulker_box\": [],\n" +
-          "\"minecraft:magenta_shulker_box\": [],\n" +
-          "\"minecraft:light_blue_shulker_box\": [],\n" +
-          "\"minecraft:yellow_shulker_box\": [],\n" +
-          "\"minecraft:lime_shulker_box\": [],\n" +
-          "\"minecraft:pink_shulker_box\": [],\n" +
-          "\"minecraft:gray_shulker_box\": [],\n" +
-          "\"minecraft:silver_shulker_box\": [],\n" +
-          "\"minecraft:cyan_shulker_box\": [],\n" +
-          "\"minecraft:purple_shulker_box\": [],\n" +
-          "\"minecraft:blue_shulker_box\": [],\n" +
-          "\"minecraft:brown_shulker_box\": [],\n" +
-          "\"minecraft:green_shulker_box\": [],\n" +
-          "\"minecraft:red_shulker_box\": [],\n" +
-          "\"minecraft:black_shulker_box\": [],\n" +
-          "\"tfc:smooth/granite\": [],\n" +
-          "\"tfc:smooth/diorite\": [],\n" +
-          "\"tfc:smooth/gabbro\": [],\n" +
-          "\"tfc:smooth/shale\": [],\n" +
-          "\"tfc:smooth/claystone\": [],\n" +
-          "\"tfc:smooth/rocksalt\": [],\n" +
-          "\"tfc:smooth/limestone\": [],\n" +
-          "\"tfc:smooth/conglomerate\": [],\n" +
-          "\"tfc:smooth/dolomite\": [],\n" +
-          "\"tfc:smooth/chert\": [],\n" +
-          "\"tfc:smooth/chalk\": [],\n" +
-          "\"tfc:smooth/rhyolite\": [],\n" +
-          "\"tfc:smooth/basalt\": [],\n" +
-          "\"tfc:smooth/andesite\": [],\n" +
-          "\"tfc:smooth/dacite\": [],\n" +
-          "\"tfc:smooth/quartzite\": [],\n" +
-          "\"tfc:smooth/slate\": [],\n" +
-          "\"tfc:smooth/phyllite\": [],\n" +
-          "\"tfc:smooth/schist\": [],\n" +
-          "\"tfc:smooth/gneiss\": [],\n" +
-          "\"tfc:smooth/marble\": [],\n" +
-          "\"tfc:stairs/smooth/andesite\": [],\n" +
-          "\"tfc:stairs/smooth/basalt\": [],\n" +
-          "\"tfc:stairs/smooth/chalk\": [],\n" +
-          "\"tfc:stairs/smooth/chert\": [],\n" +
-          "\"tfc:stairs/smooth/claystone\": [],\n" +
-          "\"tfc:stairs/smooth/conglomerate\": [],\n" +
-          "\"tfc:stairs/smooth/dacite\": [],\n" +
-          "\"tfc:stairs/smooth/diorite\": [],\n" +
-          "\"tfc:stairs/smooth/dolomite\": [],\n" +
-          "\"tfc:stairs/smooth/gabbro\": [],\n" +
-          "\"tfc:stairs/smooth/gneiss\": [],\n" +
-          "\"tfc:stairs/smooth/granite\": [],\n" +
-          "\"tfc:stairs/smooth/limestone\": [],\n" +
-          "\"tfc:stairs/smooth/marble\": [],\n" +
-          "\"tfc:stairs/smooth/phyllite\": [],\n" +
-          "\"tfc:stairs/smooth/quartzite\": [],\n" +
-          "\"tfc:stairs/smooth/rhyolite\": [],\n" +
-          "\"tfc:stairs/smooth/rocksalt\": [],\n" +
-          "\"tfc:stairs/smooth/schist\": [],\n" +
-          "\"tfc:stairs/smooth/shale\": [],\n" +
-          "\"tfc:stairs/smooth/slate\": [],\n" +
-          "\"tfc:slab/smooth/andesite\": [],\n" +
-          "\"tfc:slab/smooth/basalt\": [],\n" +
-          "\"tfc:slab/smooth/chalk\": [],\n" +
-          "\"tfc:slab/smooth/chert\": [],\n" +
-          "\"tfc:slab/smooth/claystone\": [],\n" +
-          "\"tfc:slab/smooth/conglomerate\": [],\n" +
-          "\"tfc:slab/smooth/dacite\": [],\n" +
-          "\"tfc:slab/smooth/diorite\": [],\n" +
-          "\"tfc:slab/smooth/dolomite\": [],\n" +
-          "\"tfc:slab/smooth/gabbro\": [],\n" +
-          "\"tfc:slab/smooth/gneiss\": [],\n" +
-          "\"tfc:slab/smooth/granite\": [],\n" +
-          "\"tfc:slab/smooth/limestone\": [],\n" +
-          "\"tfc:slab/smooth/marble\": [],\n" +
-          "\"tfc:slab/smooth/phyllite\": [],\n" +
-          "\"tfc:slab/smooth/quartzite\": [],\n" +
-          "\"tfc:slab/smooth/rhyolite\": [],\n" +
-          "\"tfc:slab/smooth/rocksalt\": [],\n" +
-          "\"tfc:slab/smooth/schist\": [],\n" +
-          "\"tfc:slab/smooth/shale\": [],\n" +
-          "\"tfc:slab/smooth/slate\": [],\n" +
-          "\"tfc:double_slab/smooth/andesite\": [],\n" +
-          "\"tfc:double_slab/smooth/basalt\": [],\n" +
-          "\"tfc:double_slab/smooth/chalk\": [],\n" +
-          "\"tfc:double_slab/smooth/chert\": [],\n" +
-          "\"tfc:double_slab/smooth/claystone\": [],\n" +
-          "\"tfc:double_slab/smooth/conglomerate\": [],\n" +
-          "\"tfc:double_slab/smooth/dacite\": [],\n" +
-          "\"tfc:double_slab/smooth/diorite\": [],\n" +
-          "\"tfc:double_slab/smooth/dolomite\": [],\n" +
-          "\"tfc:double_slab/smooth/gabbro\": [],\n" +
-          "\"tfc:double_slab/smooth/gneiss\": [],\n" +
-          "\"tfc:double_slab/smooth/granite\": [],\n" +
-          "\"tfc:double_slab/smooth/limestone\": [],\n" +
-          "\"tfc:double_slab/smooth/marble\": [],\n" +
-          "\"tfc:double_slab/smooth/phyllite\": [],\n" +
-          "\"tfc:double_slab/smooth/quartzite\": [],\n" +
-          "\"tfc:double_slab/smooth/rhyolite\": [],\n" +
-          "\"tfc:double_slab/smooth/rocksalt\": [],\n" +
-          "\"tfc:double_slab/smooth/schist\": [],\n" +
-          "\"tfc:double_slab/smooth/shale\": [],\n" +
-          "\"tfc:double_slab/smooth/slate\": []\n" +
-          "},\n" +
-          "\"2,2,true\": {\n" +
-          "\"minecraft:magma\": [],\n" +
-          "\"minecraft:cactus\": [],\n" +
-          "\"tfc:plants/barrel_cactus\": []\n" +
-          "},\n" +
-          "\"2,3,false\": {\n" +
-          "\"minecraft:obsidian\": [],\n" +
-          "\"minecraft:jukebox\": [],\n" +
-          "\"minecraft:gold_block\": [],\n" +
-          "\"tfc:fire_bricks\": []\n" +
-          "},\n" +
-          "\"2.5,4,false\": {\n" +
-          "\"minecraft:diamond_block\": [],\n" +
-          "\"minecraft:beacon\": []\n" +
-          "}\n" +
-          "}";
-}
